all: add local development environment
This local development environment includes a local version of the
service connected with a gremlin server and Jupyter Notebooks. It also
adds the test suite skeleton. Both, the local environment and the test
suite, are run inside Docker using docker-compose.

For the sake of demonstrating how everything fit together, this commit
also adds dummy tests and endpoints.