Add universe vertex (#5)
Adds a class with attributes that statically defines the namespace and
version of the Asset Inventory Universe. It ensures a Vertex containing the
current Universe version and namespace exists. The version of the vertex is
stored as an integer instead of a string, so it is easy the sort all the
universe vertices by version. It also ensures that when a new vertex is added to the
Inventory it is linked to the last Universe vertex.