script: improve the startup time of the local and test environments
This commit modifies the docker-compose set-up so the Gremlin server and
Jupyter Notebook environments are only launched during the first execution and
reused in the following ones. Also, it merges all the docker-compose specs in a
single file. The docs and the notebooks have been updated accordingly.

The feature consisting on using a remote Neptune server from the local
development environment has been removed. It has not shown to be useful while
adding quite a lot of complexity.